requirements.txt
# 聊天机器人项目依赖
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
